# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Construction Control System API
  description: Микросервисная система управления строительными объектами
  version: 1.0.0
  contact:
    name: ООО «СистемаКонтроля»
    email: support@systemcontrol.ru

servers:
  - url: http://localhost:5000
    description: API Gateway
  - url: http://localhost:5001  
    description: User Service
  - url: http://localhost:5002
    description: Order Service

paths:
  /users/auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /users/auth/login:
    post:
      tags:
        - Auth
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные учетные данные

  /orders:
    post:
      tags:
        - Orders
      summary: Создание заказа
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          $ref: '#/components/schemas/ApiError'

    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string

    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
          maxLength: 100

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        userId:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string

    CreateOrderRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required:
        - product
        - quantity
        - price
      properties:
        product:
          type: string
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          minimum: 0

    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, in_progress, completed, cancelled]
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time